requirements.txt
accelerate==1.2.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.7.0
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
dataclasses-json==0.6.7
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.6
filelock==3.13.1
frozenlist==1.5.0
fsspec==2024.2.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.26.5
idna==3.10
Jinja2==3.1.3
jiter==0.8.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.10
langchain-community==0.3.10
langchain-core==0.3.22
langchain-text-splitters==0.3.2
langsmith==0.1.147
MarkupSafe==2.1.5
marshmallow==3.23.1
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
neo4j==5.27.0
networkx==3.2.1
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
openai==1.57.1
orjson==3.10.12
outcome==1.3.0.post0
packaging==24.2
pillow==10.2.0
propcache==0.2.1
psutil==6.1.0
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic_core==2.27.1
PySocks==1.7.1
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.14.1
selenium==4.27.1
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.36
starlette==0.41.3
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
tqdm==4.67.1
transformers==4.47.0
trio==0.27.0
trio-websocket==0.11.1
triton==2.3.1
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0
xformers==0.0.27
yarl==1.18.3
